+++ S4-DemoApplication/src/main.tsx
@@ -3,9 +3,10 @@ import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App'
 import { AuthProvider } from './contexts/AuthContext'
-import { initSentry } from './lib/sentry'
+import { initSentry, reportWebVitals } from './lib/sentry'
 
 initSentry()
+reportWebVitals()
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>